part-1 latest version sep3

+++ src/main/java/be/intecbrussel/student/service/DaoInterface.java
@@ -5,7 +5,7 @@ import java.util.Optional;
 
 public interface DaoInterface<T>{
     Optional<T> get(String s);
-    Optional<T> get(long id);
+    Optional<T> get(Integer id);
 
     List<T> getAll();
